requirements.txt
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0
python-json-logger>=2.0.0
cryptography>=41.0.0
redis>=5.0.0
prometheus-client>=0.19.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pytest>=7.0.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0 